Move up file utilities from src/dependencies/utils to src/utils
