+++ packages/expo-modules-autolinking/src/utils.ts
@@ -1,6 +1,8 @@
 import fs from 'fs';
 import path from 'path';
 
+import { memoize } from './memoize';
+
 /** List filtered top-level files in `targetPath` (returns absolute paths) */
 export async function listFilesSorted(
   targetPath: string,
@@ -78,3 +80,34 @@ export const fileExistsAsync = async (file: string): Promise<string | null> => {
   const stat = await fs.promises.stat(file).catch(() => null);
   return stat?.isFile() ? file : null;
 };
+
+export const fastJoin: (from: string, append: string) => string =
+  path.sep === '/'
+    ? (from, append) => `${from}${path.sep}${append}`
+    : (from, append) =>
+        `${from}${path.sep}${append[0] === '@' ? append.replace('/', path.sep) : append}`;
+
+export const maybeRealpath = async (target: string): Promise<string | null> => {
+  try {
+    return await fs.promises.realpath(target);
+  } catch {
+    return null;
+  }
+};
+
+export type PackageJson = Record<string, unknown> & { name?: string; version?: string };
+
+export const loadPackageJson = memoize(async function loadPackageJson(
+  jsonPath: string
+): Promise<PackageJson | null> {
+  try {
+    const packageJsonText = await fs.promises.readFile(jsonPath, 'utf8');
+    const json = JSON.parse(packageJsonText);
+    if (typeof json !== 'object' || json == null) {
+      return null;
+    }
+    return json;
+  } catch {
+    return null;
+  }
+});
